+++ mimesis/providers/generic.py
@@ -58,7 +58,7 @@ class Generic(BaseProvider):
     )
 
     def __init__(
-            self, locale: Locale = Locale.DEFAULT, seed: Optional[Seed] = None
+        self, locale: Locale = Locale.DEFAULT, seed: Optional[Seed] = None
     ) -> None:
         """Initialize attributes lazily."""
         super().__init__(seed=seed)
@@ -131,8 +131,8 @@ class Generic(BaseProvider):
             except AttributeError:
                 name = cls.__name__.lower()
 
-            if 'seed' in kwargs:
-                kwargs.pop('seed')
+            if "seed" in kwargs:
+                kwargs.pop("seed")
 
             setattr(self, name, cls(seed=self.seed, **kwargs))
         else: